requirements.txt
flask==2.3.2
requests==2.32.0
pycryptodome==3.20.1
protobuf==4.26.3